Move some events and exception to persistence module
