+++ config/src/test/java/com/alibaba/nacos/config/server/utils/ConfigExecutorTest.java
@@ -40,21 +40,6 @@ public class ConfigExecutorTest {
     }
     
     @Test
-    public void testExecuteEmbeddedDump() throws InterruptedException {
-        
-        AtomicInteger atomicInteger = new AtomicInteger();
-        
-        Runnable runnable = atomicInteger::incrementAndGet;
-        
-        ConfigExecutor.executeEmbeddedDump(runnable);
-    
-        TimeUnit.MILLISECONDS.sleep(20);
-        
-        Assert.assertEquals(1, atomicInteger.get());
-        
-    }
-    
-    @Test
     public void testScheduleCorrectUsageTask() throws InterruptedException {
         
         AtomicInteger atomicInteger = new AtomicInteger();
@@ -92,11 +77,11 @@ public class ConfigExecutorTest {
         Runnable runnable = atomicInteger::incrementAndGet;
         
         ConfigExecutor.scheduleAsyncNotify(runnable, 20, TimeUnit.MILLISECONDS);
-    
+        
         Assert.assertEquals(0, atomicInteger.get());
         
         TimeUnit.MILLISECONDS.sleep(40);
-    
+        
         Assert.assertEquals(1, atomicInteger.get());
     }
