+++ persistence/src/main/java/com/alibaba/nacos/persistence/constants/PersistenceConstant.java
@@ -44,4 +44,11 @@ public class PersistenceConstant {
      * The derby base dir.
      */
     public static final String DERBY_BASE_DIR = "derby-data";
+    
+    /**
+     * Specifies that reads wait without timeout.
+     */
+    public static final String EXTEND_NEED_READ_UNTIL_HAVE_DATA = "00--0-read-join-0--00";
+    
+    public static final String CONFIG_MODEL_RAFT_GROUP = "nacos_config";
 }
