+++ persistence/src/main/java/com/alibaba/nacos/persistence/exception/NJdbcException.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 1999-2018 Alibaba Group Holding Ltd.
+ * Copyright 1999-2023 Alibaba Group Holding Ltd.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.alibaba.nacos.config.server.exception;
+package com.alibaba.nacos.persistence.exception;
 
 import org.springframework.dao.DataAccessException;
